I init getEndpointsWithScores lib.
