+++ hooks/useEndpointScores.tsx
@@ -1,5 +1,5 @@
-import fetchLeaderboard from '@/lib/fetchLeaderboard';
 import { endpoints as defaultEndpoints } from '@/lib/getEndpointsList';
+import getEndpointsWithScores from '@/lib/getEndpointsWithScores';
 import { useEffect, useState } from 'react';
 
 const useEndpointScores = () => {
@@ -8,23 +8,7 @@ const useEndpointScores = () => {
   useEffect(() => {
     const init = async () => {
       try {
-        const response = await fetchLeaderboard();
-        const { data } = response;
-        const stackToScoreMap = data.leaderboard.reduce((acc: any, entry: any) => {
-          const key = entry?.walletDetails?.linked_accounts?.[0]?.address?.split?.('@')?.[0];
-          console.log('key:', key);
-          const value = entry.points;
-          console.log('value:', value);
-          acc[key] = value;
-          return acc;
-        }, {});
-        console.log('stackToScoreMap:', stackToScoreMap);
-
-        const endpointsWithScores = defaultEndpoints.map((endpoint) => ({
-          ...endpoint,
-          score: stackToScoreMap[endpoint.stack] || 0,
-        }));
-
+        const endpointsWithScores = await getEndpointsWithScores();
         setEndpoints(endpointsWithScores);
       } catch (error) {
         console.error('Error fetching leaderboard data:', error);
